==> 1_station-data-cleanse.ipynb <==
# Data Cleansing for Transit Stations Database Preparation

In this notebook, we will be cleaning and standardizing the datasets for three different transit systems: Klang Valley, Montreal, and Singapore. The cleaned data will be used for the preparation of our database.

import pandas as pd
import os

data_directory = 'data'
kl_file = 'klang_valley_stations.csv'
montreal_file = 'montreal_metro.csv'
singapore_file = 'mrtsg.csv'

kl_data = pd.read_csv(os.path.join(data_directory, kl_file))
montreal_data = pd.read_csv(os.path.join(data_directory, montreal_file))
singapore_data = pd.read_csv(os.path.join(data_directory, singapore_file))



## Data Exploration
Before we start cleaning, let's explore the datasets to understand their structure and the cleaning tasks that might be necessary.

print("Klang Valley Dataset")
print(kl_data.head())
print("\nMontreal Dataset")
print(montreal_data.head())
print("\nSingapore Dataset")
print(singapore_data.head())


## Data Cleansing
The goal of the data cleansing is to ensure that each dataset has a consistent structure and standardized columns. The exact steps will depend on the initial state of the datasets.

### Klang Valley Dataset

print(kl_data.columns)
# We would now normalize/standardize the column names of the dataframe to ensure consistency
kl_data.columns = kl_data.columns.str.lower().str.replace(' ', '_')

#ensure only first letter of a names is capitalized
kl_data['name'] = kl_data['name'].str.title()


# Rename columns
kl_data = kl_data.rename(columns={
    'city': 'region',
    'stop_id':'station_code'
})

# Rename rows

kl_data['name'] = kl_data['name'].replace({
    "Kl Sentral": "KL Sentral",
    "Ukm": "UKM",
    "Pwtc": "PWTC",
    "Taman Perindustrian Puchong (Tpp)":"Taman Perindustrian Puchong (TPP)",
    "Klcc":"KLCC",
    "Ss15":"SS15",
    "Ss18": "SS18", 
    "Usj7" : "USJ7",
    "Usj21" : "USJ21",
    "Klia2" : "KLIA2",
    "Ttdi" : "TTDI",
    "Sunu Monash" : "SunU Monash",
    "South Quay-Usj1" : "South Quay-USJ1",
    "Persiaran Klcc" : "Persiaran KLCC",
    "Upm" : "UPM",
    "Bu11" : "BU11",
    "Ss7" : "SS7",
    "Uitm" : "UiTM"
            })


print(kl_data.columns)
kl_data.head()

### Montreal Dataset

print(montreal_data.columns)
# We would now normalize/standardize the column names of the dataframe to ensure consistency
montreal_data.columns = montreal_data.columns.str.lower().str.replace(' ', '_')

#drop irrelevant columns
montreal_data=montreal_data.drop(columns=['index'])


# Rename columns
montreal_data = montreal_data.rename(columns={
    '_stop_id': 'station_code',
    'city':'region'
})



print(montreal_data.columns)
montreal_data.head()



### Singapore Dataset

print(singapore_data.columns)
# We would now normalize/standardize the column names of the dataframe to ensure consistency
singapore_data.columns = singapore_data.columns.str.lower().str.replace(' ', '_')

#drop irrelevant columns
singapore_data=singapore_data.drop(columns=['objectid','x','y'])

#ensure only first letter of a names is capitalized
singapore_data['line_colour'] = singapore_data['line_colour'].str.title()

# Rename columns
singapore_data = singapore_data.rename(columns={
    'stn_no': 'station_code',
    'route_code':'route_id',
    'city':'region'
})



print(singapore_data.columns)
singapore_data.head()

## Save Cleaned Datasets
Finally, we will save the cleaned datasets to new CSV files for use in creating our database

# Define the directory where you want to save the cleaned data
cleansed_data_directory = 'data_cleansed'
cleansed_kl_file = 'klang_valley_stations_cleansed.csv'
cleansed_montreal_file = 'montreal_stations_cleansed.csv'
cleansed_singapore_file = 'singapore_stations_cleansed.csv'

# Save the cleaned dataframes
kl_data.to_csv(os.path.join(cleansed_data_directory, cleansed_kl_file), index=False)
montreal_data.to_csv(os.path.join(cleansed_data_directory, cleansed_montreal_file), index=False)
singapore_data.to_csv(os.path.join(cleansed_data_directory, cleansed_singapore_file), index=False)
