==> 6_supabase_test.ipynb <==
# Testing Out Supabase Client

!echo $SUPABASE_URL

import os
import pandas as pd
from dotenv import load_dotenv
load_dotenv()

from supabase import create_client, Client

url: str = os.environ.get("SUPABASE_URL")

key: str = os.environ.get("SUPABASE_SERVICE_KEY")
supabase: Client = create_client(url, key)


response = supabase.table('station_isochrones').select("*").execute()
data,_ = response
data[1]
df = pd.DataFrame(data[1])
df
